requirements.txt
# GovProposal Portal — Dependencies
# CUI // SP-PROPIN

# Web framework
flask>=3.0
jinja2>=3.1

# Data & search
pyyaml>=6.0
numpy>=1.24
rank_bm25>=0.2

# Document processing
python-docx>=1.0           # Word document generation
pypdf>=4.0                 # PDF parsing (solicitations)
openpyxl>=3.1              # Excel for pricing volumes

# HTTP
requests>=2.31

# LLM (optional — Bedrock via boto3 or OpenAI-compat)
boto3>=1.34                # AWS Bedrock
openai>=1.0                # OpenAI-compatible providers

# Testing
pytest>=8.0
pytest-cov>=5.0

# Security
python-dotenv>=1.0         # .env loading
